Fix Remove api.go file. pre-compute URL + checksum and replace api_endpoint parameter by optional path_prefix
